pyproject.toml
[tool.poetry]
name = "pythonproject1"
version = "0.1.0"
description = ""
authors = ["arystambek-dimash <arystambekdimash005@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
fastapi = "^0.112.1"
pydantic = {extras = ["email"], version = "^2.8.2"}
pydantic-settings = "^2.4.0"
sqlalchemy = {extras = ["asyncio"], version = "^2.0.32"}
uvicorn = "^0.30.6"
psycopg2-binary = "^2.9.9"
asyncpg = "^0.29.0"
pyjwt = "^2.9.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
boto3 = "^1.35.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
